Library

In the context of the package, exponential family distributions are represented in the form $\begin{aligned} f_X(x\mid\theta) = h(x)\,\exp\!\bigl[\,\eta(\theta) \cdot T(x) - A(\theta)\,\bigr] \end{aligned}$ where $h(x)$ - basemeasure, $T(x)$ - sufficient statistics, $A(\theta)$ - log partition, $\eta(\theta)$ - natural parameters.

Main Structure

ExponentialFamily.ExponentialFamilyDistribution — Type

ExponentialFamilyDistribution(::Type{T}, naturalparameters, [ conditioner, basemeasure, sufficientstatistics, logpartition, support, supportcheck ])

ExponentialFamilyDistribution structure represents a generic exponential family distribution in natural parameterization. Methods defined are basemeasure ,sufficientstatistics, naturalparameters, logpartition and support.

getbasemeasure returns a positive a valued function.

getsufficientstatistics returns an iterable of functions such as [x, x^2] or [x, logx].

getnaturalparameters returns an iterable holding the values of the natural parameters.

getlogpartition return a function that depends on the naturalparameters and it ensures that the distribution is normalized to 1.

support returns the set that the distribution is defined over. Could be real numbers, positive integers, 3d cube etc. Use the insupport to check if a values is in support.

See also: getbasemeasure, getsufficientstatistics, getnaturalparameters, getlogpartition, getsupport

Methods

ExponentialFamily.getlogpartition — Function

getlogpartition(ef::ExponentialFamilyDistribution)

Get the log partition of the exponential family distribution ef.

ExponentialFamily.getbasemeasure — Function

getbasemeasure(ef::ExponentialFamilyDistribution)

Get the base measure of the exponential family distribution ef.

ExponentialFamily.getsufficientstatistics — Function

getsufficientstatistics(ef::ExponentialFamilyDistribution)

Get the sufficient statistics of the exponential family distribution ef.

ExponentialFamily.getnaturalparameters — Function

getnaturalparameters(exponentialfamily::ExponentialFamilyDistribution)

Get the natural parameters of the exponential family distribution ef.

ExponentialFamily.getsupport — Function

getsupport(ef::ExponentialFamilyDistribution)

Get the support of the exponential family distribution ef.

Distributions.logpdf — Function

Distributions.logpdf(exponentialfamily::ExponentialFamilyDistribution{T}, x) where {T <: Distribution}

Evaluate the log-density of the exponential family distribution for the input x with constant base measure.

Arguments

exponentialfamily: The exponential family distribution.

x: The input value.

Returns

The log-density of exponentialfamily evaluated at x with constant base measure.

Distributions.logpdf(exponentialfamily::ExponentialFamilyDistribution{T, H, S, P, C, Z, A, B}, x, ::ConstantBaseMeasure)

Evaluate the log-density of the exponential family distribution for the input x with constant base measure.

Arguments

exponentialfamily: The exponential family distribution.

x: The input value.

::ConstantBaseMeasure: A marker for the constant base measure (used for dispatch).

Returns

The log-density of exponentialfamily evaluated at x with constant base measure.

Distributions.logpdf(exponentialfamily::ExponentialFamilyDistribution{T, H, S, P, C, Z, A, B}, x, ::NonConstantBaseMeasure)

Evaluate the log-density of the exponential family distribution for the input x with non-constant base measure.

Arguments

exponentialfamily: The exponential family distribution.

x: The input value.

::NonConstantBaseMeasure: A marker for the non-constant base measure (used for dispatch).

Returns

The log-density of exponentialfamily evaluated at x with non-constant base measure.

Distributions.logpdf(exponentialfamily::ExponentialFamilyDistribution{T, H, S, P, C, Z, A, B}, x)

Evaluate the log-density of the exponential family distribution for the input x with constant base measure.

Arguments

exponentialfamily: The exponential family distribution.

x: The input value.

Returns

The log-density of exponentialfamily evaluated at x with constant base measure.

Distributions.pdf — Function

Distributions.pdf(exponentialfamily::ExponentialFamilyDistribution, x)

Evaluate the probability density function of the exponential family distribution for the input x.

ExponentialFamily.fisherinformation — Function

The function fisherinformation is used to compute the Fisher information matrix of an exponential family distribution. fisherinformation calculates the Fisher information matrix based on the natural parameters of the distribution.

Additional distributions

These are the distributions that are not included in the Distributions.jl package.

ExponentialFamily.Contingency — Type

Contingency(P, renormalize = Val(true))

The contingency distribution is a multivariate generalization of the categorical distribution. As a bivariate distribution, the contingency distribution defines the joint probability over two unit vectors v1 and v2 with one hot encoding. Or it can be thought of as the joint distribution of two categoricals with supports v1 ={ 1,2,...,N} and v2 ={ 1,2,...,N}. The parameter P encodes a contingency matrix that specifies the probability of co-occurrence.

v1 ∈ {0, 1}^d1 where Σ_j v1_j = 1
v2 ∈ {0, 1}^d2 where Σ_k v2_k = 1

P ∈ [0, 1]^{d1 × d2}, where Σ_jk P_jk = 1

f(v1, v2, P) = Contingency(out1, out2 | P) = Π_jk P_jk^{v1_j * v2_k}

A Contingency distribution over more than two variables requires higher-order tensors as parameters; these are not implemented in ExponentialFamily.jl yet.

Fields:

P, required, contingency matrix

renormalize, optional, supports either Val(true) or Val(false), specifies whether matrix P must be automatically renormalized. Does not modify the original P and allocates a new one for the renormalized version. If set to false the contingency matrix P must be normalized by hand, otherwise the result of related calculations might be wrong

ExponentialFamily.ContinuousBernoulli — Type

ContinuousBernoulli{T}

A univariate continuous Bernoulli distribution parametrized by its success probability λ.

Fields

λ: The success probability of the continuous Bernoulli distribution. It should be a real number in the interval (0, 1)

ExponentialFamily.MatrixDirichlet — Type

MatrixDirichlet{T <: Real, A <: AbstractMatrix{T}} <: ContinuousMatrixDistribution

A matrix-valued Dirichlet distribution, where T is the element type of the matrix A. The a field stores the matrix parameter of the distribution.

Fields

a::A: The matrix parameter of the Dirichlet distribution.

ExponentialFamily.GammaShapeRate — Type

GammaShapeRate{T <: Real}

A univariate gamma distribution parametrized by its shape a and rate b.

Fields

a: The shape parameter of the gamma distribution. It should be a positive real number.

b: The rate parameter of the gamma distribution. It should be a positive real number.

ExponentialFamily.GammaShapeScale — Type

GammaShapeScale{T}

A continuous univariate gamma distribution parametrized by its shape α and scale β parameters.

Fields

α: The shape parameter of the gamma distribution. It should be a positive real number.

β: The scale parameter of the gamma distribution. It should be a positive real number.

Note

GammaShapeScale is an alias for Gamma from Distributions.jl.

ExponentialFamily.NormalMeanPrecision — Type

NormalMeanPrecision{T <: Real} <: ContinuousUnivariateDistribution

A normal distribution with a known mean μ and precision w.

Fields

μ::T: The mean of the normal distribution.

w::T: The precision of the normal distribution.

ExponentialFamily.NormalMeanVariance — Type

NormalMeanVariance{T <: Real} <: ContinuousUnivariateDistribution

A normal distribution with a known mean μ and variance v.

Fields

μ::T: The mean of the normal distribution.

v::T: The variance of the normal distribution.

ExponentialFamily.NormalWeightedMeanPrecision — Type

NormalWeightedMeanPrecision{T <: Real} <: ContinuousUnivariateDistribution

A normal distribution parametrized by its natural parameters: the weighted mean xi and precision w.

Fields

xi::T: The weighted mean of the normal distribution. xi is computed as w * μ, where μ is the mean of the distribution.

w::T: The precision (inverse variance) of the normal distribution.

ExponentialFamily.MvNormalMeanPrecision — Type

MvNormalMeanPrecision{T <: Real, M <: AbstractVector{T}, P <: AbstractMatrix{T}} <: AbstractMvNormal

A multivariate normal distribution with mean μ and precision matrix Λ, where T is the element type of the vectors M and matrices P.

Fields

μ::M: The mean vector of the multivariate normal distribution.

Λ::P: The precision matrix (inverse of the covariance matrix) of the multivariate normal distribution.

ExponentialFamily.MvNormalMeanCovariance — Type

MvNormalMeanCovariance{T <: Real, M <: AbstractVector{T}, P <: AbstractMatrix{T}} <: AbstractMvNormal

A multivariate normal distribution with mean μ and covariance matrix Σ, where T is the element type of the vectors M and matrices P.

Fields

μ::M: The mean vector of the multivariate normal distribution.

Σ::P: The covariance matrix of the multivariate normal distribution

ExponentialFamily.MvNormalWeightedMeanPrecision — Type

MvNormalWeightedMeanPrecision{T <: Real, M <: AbstractVector{T}, P <: AbstractMatrix{T}} <: AbstractMvNormal

A multivariate normal distribution with a weighted mean vector xi and precision matrix Λ, where T is the element type of the vectors M and matrices P. This struct represents a natural parametrization of a multivariate Gaussian distribution.

Fields

xi::M: The weighted mean vector of the multivariate normal distribution.

Λ::P: The precision matrix (inverse of the covariance matrix) of the multivariate normal distribution.

ExponentialFamily.JointNormal — Type

JointNormal{D, S}

JointNormal is an auxilary structure used for the joint marginal over Normally distributed variables. JointNormal stores a vector with the original dimensionalities (ds), so statistics can later be re-separated.

Fields

dist: joint distribution (typically just a big MvNormal distribution, but maybe a tuple of individual means and covariance matrices)

ds: a tuple with the original dimensionalities of individual Normal distributions

ds[k] = (n,) where n is an integer indicates Multivariate normal of size n

ds[k] = () indicates Univariate normal

ExponentialFamily.WishartFast — Type

WishartFast{T <: Real, A <: AbstractMatrix{T}} <: ContinuousMatrixDistribution

The WishartFast struct represents a fast version of the Wishart distribution. It is similar to the Wishart distribution from Distributions.jl, but it does not check input arguments, allowing the creation of improper Wishart messages.

For model creation and regular usage, it is recommended to use Wishart from Distributions.jl. The WishartFast distribution is intended for internal purposes and should not be directly used by regular users.

Fields

ν::T: The degrees of freedom parameter of the Wishart distribution.

invS::A: The inverse scale matrix parameter of the Wishart distribution.

Note

Internally, WishartFast stores and creates the inverse of its scale matrix. However, the params() function returns the scale matrix itself for backward compatibility. This is done to ensure better stability in the message passing update rules for ReactiveMP.jl.

ExponentialFamily.InverseWishartFast — Type

InverseWishartFast{T <: Real, A <: AbstractMatrix{T}} <: ContinuousMatrixDistribution

The InverseWishartFast struct represents an improper Inverse Wishart distribution. It is similar to the InverseWishart distribution from Distributions.jl, but it does not check input arguments, allowing the creation of improper InverseWishart messages.

For model creation and regular usage, it is recommended to use InverseWishart from Distributions.jl. The InverseWishartFast distribution is intended for internal purposes and should not be directly used by regular users.

Fields

ν::T: The degrees of freedom parameter of the inverse Wishart distribution.

S::A: The scale matrix parameter of the inverse Wishart distribution.

Note

The InverseWishartFast distribution does not enforce input argument validation, making it suitable for specialized cases where improper message constructions are needed.

Prod related methods

Base.prod — Method

prod(strategy, left, right)

prod function is used to find a product of two probability distrubutions (or any other objects) over same variable (e.g. 𝓝(x|μ1, σ1) × 𝓝(x|μ2, σ2)). There are multiple strategies for prod function, e.g. ClosedProd, ProdGeneric or ProdPreserveType.

Examples:

using ExponentialFamily

product = prod(ClosedProd(), NormalMeanVariance(-1.0, 1.0), NormalMeanVariance(1.0, 1.0))

mean(product), var(product)

# output
(0.0, 0.5)

See also: closed_prod_rule, ClosedProd, ProdGeneric

ExponentialFamily.ProdPreserveType — Type

ProdPreserveType{T}

ProdPreserveType is one of the strategies for prod function. This strategy constraint an output of a prod to be in some specific form. By default it fallbacks to a ClosedProd strategy and converts an output to a prespecified type but can be overwritten for some distributions for better performance. See also: prod, ClosedProd, ProdPreserveTypeLeft, ProdPreserveTypeRight

ExponentialFamily.ProdPreserveTypeLeft — Type

ProdPreserveTypeLeft

ProdPreserveTypeLeft is one of the strategies for prod function. This strategy constraint an output of a prod to be in the functional form as left argument. By default it fallbacks to a ProdPreserveType strategy and converts an output to a prespecified type but can be overwritten for some distributions for better performance. See also: prod, ProdPreserveType, ProdPreserveTypeRight

ExponentialFamily.ProdGeneric — Type

ProdGeneric{C}

ProdGeneric is one of the strategies for prod function. This strategy does not fail in case of no closed rule is available, but simply creates a product tree, there all nodes represent the prod function and all leaves are valid Distribution object. This object does not define any statistical properties (such as mean or var etc) and cannot be used during the inference procedure. However this object plays imporant part in the functional form constraints implementation. In a few words this object keeps all the information of a product of messages and propagates this information in the functional form constraint.

ProdGeneric has a "fallback" method, which it may or may not use under some circumstances. For example if the fallback method is ClosedProd (which is the default one) - ProdGeneric will try to optimize prod tree with analytical solutions where possible.

See also: prod, ProductDistribution, ClosedProd, ProdPreserveType, closed_prod_rule, LinearizedProductDistribution

ExponentialFamily.ProdPreserveTypeRight — Type

ProdPreserveTypeRight

ProdPreserveTypeRight is one of the strategies for prod function. This strategy constraint an output of a prod to be in the functional form as right argument. By default it fallbacks to a ProdPreserveType strategy and converts an output to a prespecified type but can be overwritten for some distributions for better performance. See also: prod, ProdPreserveType, ProdPreserveTypeLeft

ExponentialFamily.closed_prod_rule — Function

closed_prod_rule(::Type, ::Type)

Returns either ClosedProd or ClosedProdUnknown for two given distribution types. Returns ClosedProdUnknown by default. See also: prod, ClosedProd, ProdGeneric

ExponentialFamily.ClosedProd — Type

ClosedProd

ClosedProd is one of the strategies for prod function. This strategy uses analytical prod methods but does not constraint a prod to be in any specific form. It throws an NoClosedProdException if no analytical rules is available, use ProdGeneric prod strategy to fallback to approximation methods.

Note: ClosedProd ignores missing values and simply returns the non-missing argument. Returns missing in case if both arguments are missing.

See also: prod, ProdPreserveType, ProdGeneric

ExponentialFamily.LinearizedProductDistribution — Type

LinearizedProductDistribution

An efficient linearized implementation of product of multiple generic ExponentialFamilyDistribution objects. This structure prevents ProductDistribution tree from growing too much in case of identical objects. This trick significantly reduces Julia compilation times when closed product rules are not available but distributions are of the same type. Essentially this structure linearizes leaves of the ProductDistribution tree in case if it sees objects of the same type (via dispatch).

See also: ProductDistribution

ExponentialFamily.ProductDistribution — Type

ProductDistribution{L, R}(left::L, right::R)

A product distribution structure composed of two distributions left and right.

Arguments

left: The left component distribution.

right: The right component distribution.