Library API

Product API

One of the central functions in this package is the ability to calculate the product of two distributions over the same variable. You can also refer to the corresponding example for practical usage.

The prod function is a key feature of this package. It accepts a strategy as its first argument, which defines how the prod function should behave and what results you can expect.

Base.prod — Method

prod(strategy, left, right)

prod function is used to find a product of two probability distributions (or any other objects) over same variable (e.g. 𝓝(x|μ1, σ1) × 𝓝(x|μ2, σ2)). There are multiple strategies for prod function, e.g. ClosedProd, GenericProd or PreserveTypeProd.

Examples:

julia> product = prod(PreserveTypeProd(Distribution), NormalMeanVariance(-1.0, 1.0), NormalMeanVariance(1.0, 1.0))
NormalWeightedMeanPrecision{Float64}(xi=0.0, w=2.0)

julia> mean(product), var(product)
(0.0, 0.5)

julia> product = prod(PreserveTypeProd(NormalMeanVariance), NormalMeanVariance(-1.0, 1.0), NormalMeanVariance(1.0, 1.0))
NormalMeanVariance{Float64}(μ=0.0, v=0.5)

julia> mean(product), var(product)
(0.0, 0.5)

julia> product = prod(PreserveTypeProd(ExponentialFamilyDistribution), NormalMeanVariance(-1.0, 1.0), NormalMeanVariance(1.0, 1.0))
ExponentialFamily(NormalMeanVariance)

julia> mean(product), var(product)
(0.0, 0.5)

See also: default_prod_rule, ClosedProd, PreserveTypeProd, GenericProd

ExponentialFamily.default_prod_rule — Function

default_prod_rule(::Type, ::Type)

Returns the most suitable prod rule for two given distribution types. Returns UnspecifiedProd by default.

See also: prod, ClosedProd, GenericProd

Product strategies

For certain distributions, it's possible to compute the product using a straightforward mathematical equation, yielding a closed-form solution. However, for some distributions, finding a closed-form solution might not be feasible. Various strategies ensure consistent behavior in these situations. These strategies can either guarantee a fast and closed-form solution or, when necessary, fall back to a slower but more generic method.

ExponentialFamily.UnspecifiedProd — Type

UnspecifiedProd

A strategy for the prod function, which does not compute the prod, but instead fails in run-time and prints a descriptive error message.

See also: prod, ClosedProd, GenericProd

ExponentialFamily.ClosedProd — Type

ClosedProd

ClosedProd is one of the strategies for prod function. This strategy uses either PreserveTypeProd(Distribution) or PreserveTypeProd(ExponentialFamilyDistribution), depending on the types of the input arguments. For example, if both inputs are of type Distribution, then ClosedProd would fallback to PreserveTypeProd(Distribution).

See also: prod, PreserveTypeProd, GenericProd

ExponentialFamily.PreserveTypeProd — Type

PreserveTypeProd{T}

PreserveTypeProd is one of the strategies for prod function. This strategy constraint an output of a prod to be in some specific form. By default it uses the strategy from default_prod_rule and converts the output to the prespecified type but can be overwritten for some distributions for better performance.

julia> product = prod(PreserveTypeProd(NormalMeanVariance), NormalMeanVariance(-1.0, 1.0), NormalMeanVariance(1.0, 1.0))
NormalMeanVariance{Float64}(μ=0.0, v=0.5)

julia> mean(product), var(product)
(0.0, 0.5)

See also: prod, ClosedProd, PreserveTypeLeftProd, PreserveTypeRightProd, GenericProd

ExponentialFamily.PreserveTypeLeftProd — Type

PreserveTypeLeftProd

An alias for the PreserveTypeProd(L) where L is the type of the left argument of the prod function.

julia> product = prod(PreserveTypeLeftProd(), NormalMeanVariance(-1.0, 1.0), NormalMeanPrecision(1.0, 1.0))
NormalMeanVariance{Float64}(μ=0.0, v=0.5)

julia> mean(product), var(product)
(0.0, 0.5)

See also: prod, PreserveTypeProd, PreserveTypeRightProd, GenericProd

ExponentialFamily.PreserveTypeRightProd — Type

PreserveTypeRightProd

An alias for the PreserveTypeProd(R) where R is the type of the right argument of the prod function.

julia> product = prod(PreserveTypeRightProd(), NormalMeanVariance(-1.0, 1.0), NormalMeanPrecision(1.0, 1.0))
NormalMeanPrecision{Float64}(μ=0.0, w=2.0)

julia> mean(product), var(product)
(0.0, 0.5)

See also: prod, PreserveTypeProd, PreserveTypeLeftProd, GenericProd

ExponentialFamily.GenericProd — Type

GenericProd

GenericProd is one of the strategies for prod function. This strategy does always produces a result, even if the closed form product is not availble, in which case simply returns the ProductOf object. GenericProd sometimes fallbacks to the default_prod_rule which it may or may not use under some circumstances. For example if the default_prod_rule is ClosedProd - GenericProd will try to optimize the tree with analytical closed solutions (if possible).

See also: prod, ProductOf, ClosedProd, PreserveTypeProd, default_prod_rule

ExponentialFamily.ProductOf — Type

ProductOf

A generic structure representing a product of two distributions. Can be viewed as a tuple of (left, right). Does not check nor supports neither variate forms during the creation stage. Uses the fuse_support function to fuse supports of two different distributions.

This object does not define any statistical properties (such as mean or var etc) and cannot be used as a distribution explicitly. Instead, it must be further approximated as a member of some other distribution.

See also: prod, GenericProd, ExponentialFamily.fuse_supports

ExponentialFamily.LinearizedProductOf — Type

LinearizedProductOf

An efficient linearized implementation of product of multiple distributions. This structure prevents ProductOf tree from growing too much in case of identical objects. This trick significantly reduces Julia compilation times when closed product rules are not available but distributions are of the same type. Essentially this structure linearizes leaves of the ProductOf tree in case if it sees objects of the same type (via dispatch).

See also: ProductOf, [GenericProd]

These strategies offer flexibility and reliability when working with different types of distributions, ensuring that the package can handle a wide range of cases effectively.

Additional distributions

These are the distributions that are not included in the Distributions.jl package.

ExponentialFamily.MatrixDirichlet — Type

MatrixDirichlet{T <: Real, A <: AbstractMatrix{T}} <: ContinuousMatrixDistribution

A matrix-valued MatrixDirichlet distribution, where T is the element type of the matrix A. The a field stores the matrix parameter of the distribution.

Fields

a::A: The matrix parameter of the MatrixDirichlet distribution.

ExponentialFamily.GammaShapeRate — Type

GammaShapeRate{T <: Real}

A univariate gamma distribution parametrized by its shape a and rate b.

Fields

a: The shape parameter of the gamma distribution. It should be a positive real number.

b: The rate parameter of the gamma distribution. It should be a positive real number.

ExponentialFamily.GammaShapeScale — Type

GammaShapeScale{T}

A continuous univariate gamma distribution parametrized by its shape α and scale β parameters.

Fields

α: The shape parameter of the gamma distribution. It should be a positive real number.

β: The scale parameter of the gamma distribution. It should be a positive real number.

Note

GammaShapeScale is an alias for Gamma from Distributions.jl.

ExponentialFamily.NormalMeanPrecision — Type

NormalMeanPrecision{T <: Real} <: ContinuousUnivariateDistribution

A normal distribution with a known mean μ and precision w.

Fields

μ::T: The mean of the normal distribution.

w::T: The precision of the normal distribution.

ExponentialFamily.NormalMeanVariance — Type

NormalMeanVariance{T <: Real} <: ContinuousUnivariateDistribution

A normal distribution with a known mean μ and variance v.

Fields

μ::T: The mean of the normal distribution.

v::T: The variance of the normal distribution.

ExponentialFamily.NormalWeightedMeanPrecision — Type

NormalWeightedMeanPrecision{T <: Real} <: ContinuousUnivariateDistribution

A normal distribution parametrized by its natural parameters: the weighted mean xi and precision w.

Fields

xi::T: The weighted mean of the normal distribution. xi is computed as w * μ, where μ is the mean of the distribution.

w::T: The precision (inverse variance) of the normal distribution.

ExponentialFamily.MvNormalMeanPrecision — Type

MvNormalMeanPrecision{T <: Real, M <: AbstractVector{T}, P <: AbstractMatrix{T}} <: AbstractMvNormal

A multivariate normal distribution with mean μ and precision matrix Λ, where T is the element type of the vectors M and matrices P.

Fields

μ::M: The mean vector of the multivariate normal distribution.

Λ::P: The precision matrix (inverse of the covariance matrix) of the multivariate normal distribution.

ExponentialFamily.MvNormalMeanCovariance — Type

MvNormalMeanCovariance{T <: Real, M <: AbstractVector{T}, P <: AbstractMatrix{T}} <: AbstractMvNormal

A multivariate normal distribution with mean μ and covariance matrix Σ, where T is the element type of the vectors M and matrices P.

Fields

μ::M: The mean vector of the multivariate normal distribution.

Σ::P: The covariance matrix of the multivariate normal distribution

ExponentialFamily.MvNormalWeightedMeanPrecision — Type

MvNormalWeightedMeanPrecision{T <: Real, M <: AbstractVector{T}, P <: AbstractMatrix{T}} <: AbstractMvNormal

A multivariate normal distribution with a weighted mean vector xi and precision matrix Λ, where T is the element type of the vectors M and matrices P. This struct represents a natural parametrization of a multivariate Gaussian distribution.

Fields

xi::M: The weighted mean vector of the multivariate normal distribution.

Λ::P: The precision matrix (inverse of the covariance matrix) of the multivariate normal distribution.

ExponentialFamily.JointNormal — Type

JointNormal{D, S}

JointNormal is an auxilary structure used for the joint marginal over Normally distributed variables. JointNormal stores a vector with the original dimensionalities (ds), so statistics can later be re-separated.

Use ExponentialFamily.getcomponent(joint, index) to get a specific component of the joint distribution.

Fields

dist: joint distribution (typically just a big MvNormal distribution, but maybe a tuple of individual means and covariance matrices)

ds: a tuple with the original dimensionalities of individual Normal distributions

ds[k] = (n,) where n is an integer indicates Multivariate normal of size n

ds[k] = () indicates Univariate normal

ExponentialFamily.WishartFast — Type

WishartFast{T <: Real, A <: AbstractMatrix{T}} <: ContinuousMatrixDistribution

The WishartFast struct represents a fast version of the Wishart distribution. It is similar to the Wishart distribution from Distributions.jl, but it does not check input arguments, allowing the creation of improper Wishart messages.

For model creation and regular usage, it is recommended to use Wishart from Distributions.jl. The WishartFast distribution is intended for internal purposes and should not be directly used by regular users.

Fields

ν::T: The degrees of freedom parameter of the Wishart distribution.

invS::A: The inverse scale matrix parameter of the Wishart distribution.

Note

Internally, WishartFast stores and creates the inverse of its scale matrix. However, the params() function returns the scale matrix itself for backward compatibility. This is done to ensure better stability in the message passing update rules for ReactiveMP.jl.

ExponentialFamily.InverseWishartFast — Type

InverseWishartFast{T <: Real, A <: AbstractMatrix{T}} <: ContinuousMatrixDistribution

The InverseWishartFast struct represents an improper Inverse Wishart distribution. It is similar to the InverseWishart distribution from Distributions.jl, but it does not check input arguments, allowing the creation of improper InverseWishart messages.

For model creation and regular usage, it is recommended to use InverseWishart from Distributions.jl. The InverseWishartFast distribution is intended for internal purposes and should not be directly used by regular users.

Fields

ν::T: The degrees of freedom parameter of the inverse Wishart distribution.

S::A: The scale matrix parameter of the inverse Wishart distribution.

Note

The InverseWishartFast distribution does not enforce input argument validation, making it suitable for specialized cases where improper message constructions are needed.

ExponentialFamily.JointGaussian — Type

An alias for the JointNormal.

ExponentialFamily.NormalGamma — Type

NormalGamma{T <: Real} <: ContinuousMultivariateDistribution

A normal-gamma distribution, where T is a real number. This distribution is a joint distribution of a normal random variable with mean μ and precision λ, and a gamma-distributed random variable with shape α and rate β.

Fields

μ::T: The mean of the normal distribution.

λ::T: The precision of the normal distribution.

α::T: The shape parameter of the gamma distribution.

β::T: The rate parameter of the gamma distribution.

ExponentialFamily.MvNormalWishart — Type

MvNormalWishart{T, M <: AbstractArray{T}, V <: AbstractMatrix{T}, K <: Real, N <: Real} <: ContinuousMatrixDistribution

A multivariate normal-Wishart distribution, where T is the element type of the arrays M and matrices V, and K and N are real numbers. This distribution is a joint distribution of a multivariate normal random variable with mean μ and a Wishart-distributed random matrix with scale matrix Ψ, degrees of freedom ν, and the scalar κ as a scaling parameter.

Fields

μ::M: The mean vector of the multivariate normal distribution.

Ψ::V: The scale matrix of the Wishart distribution.

κ::K: The scaling parameter of the Wishart distribution.

ν::N: The degrees of freedom of the Wishart distribution

ExponentialFamily.FactorizedJoint — Type

FactorizedJoint

FactorizedJoint represents a joint distribution of independent random variables. Use getindex() function or square-brackets indexing to access the marginal distribution for individual variables.

Promotion type utilities

ExponentialFamily.paramfloattype — Function

paramfloattype(distribution)

Returns the underlying float type of distribution's parameters.

See also: ExponentialFamily.promote_paramfloattype, ExponentialFamily.convert_paramfloattype

ExponentialFamily.sampletype — Function

sampletype(distribution)

Returns a type of the distribution. By default fallbacks to the eltype.

See also: ExponentialFamily.samplefloattype, ExponentialFamily.promote_sampletype, ExponentialFamily.promote_samplefloattype

ExponentialFamily.samplefloattype — Function

samplefloattype(distribution)

Returns a type of the distribution or the underlying float type in case if sample is Multivariate or Matrixvariate. By default fallbacks to the deep_eltype(sampletype(distribution)).

See also: ExponentialFamily.sampletype, ExponentialFamily.promote_sampletype, ExponentialFamily.promote_samplefloattype

ExponentialFamily.promote_variate_type — Function

promote_variate_PromoteTypeConverter(::Type{ <: VariateForm }, distribution_type)

Promotes (if possible) a distribution_type to be of the specified variate form.

ExponentialFamily.promote_paramfloattype — Function

promote_paramfloattype(distributions...)

Promotes paramfloattype of the distributions to a single type. See also promote_type.

See also: ExponentialFamily.paramfloattype, ExponentialFamily.convert_paramfloattype

ExponentialFamily.promote_sampletype — Function

promote_sampletype(distributions...)

Promotes sampletype of the distributions to a single type. See also promote_type.

See also: ExponentialFamily.sampletype, ExponentialFamily.samplefloattype, ExponentialFamily.promote_samplefloattype

ExponentialFamily.promote_samplefloattype — Function

promote_samplefloattype(distributions...)

Promotes samplefloattype of the distributions to a single type. See also promote_type.

See also: ExponentialFamily.sampletype, ExponentialFamily.samplefloattype, ExponentialFamily.promote_sampletype

ExponentialFamily.convert_paramfloattype — Function

convert_paramfloattype(::Type{T}, distribution)

Converts (if possible) the params float type of the distribution to be of type T.

See also: ExponentialFamily.paramfloattype, ExponentialFamily.promote_paramfloattype

convert_paramfloattype(::Type{T}, container)

Converts (if possible) the elements of the container to be of type T.

Extra stats functions

ExponentialFamily.mirrorlog — Function

mirrorlog(x)

Returns log(1 - x).

ExponentialFamily.xtlog — Function

xtlog(x)

Returns x * log(x).

ExponentialFamily.logmvbeta — Function

logmvbeta(x)

Uses the numerically stable algorithm to compute the logarithm of the multivariate beta distribution over with the parameter vector x.

ExponentialFamily.clamplog — Function

clamplog(x)

Same as log but clamps the input argument x to be in the range tiny <= x <= typemax(x) such that log(0) does not explode.

ExponentialFamily.mvtrigamma — Function

mvtrigamma(p, x)

Computes multivariate trigamma function .

Helper utilities

ExponentialFamily.vague — Function

vague(distribution_type, [ dims... ])

vague function returns uninformative probability distribution of a given type.

ExponentialFamily.logpdf_sample_optimized — Function

logpdf_sample_optimized(distribution)

logpdf_sample_optimized function takes as an input a distribution and returns corresponding optimized two versions for taking logpdf() and sampling with rand! respectively. By default returns the same distribution, but some distributions may override default behaviour for better efficiency.

Example

julia> d = vague(MvNormalMeanPrecision, 2)
MvNormalMeanPrecision(
μ: [0.0, 0.0]
Λ: [1.0e-12 0.0; 0.0 1.0e-12]
)

julia> ExponentialFamily.logpdf_sample_optimized(d)
(FullNormal(
dim: 2
μ: [0.0, 0.0]
Σ: [1.0e12 -0.0; -0.0 1.0e12]
)
, FullNormal(
dim: 2
μ: [0.0, 0.0]
Σ: [1.0e12 -0.0; -0.0 1.0e12]
)
)

ExponentialFamily.fuse_supports — Function

fuse_supports(left, right)

Fuse supports of two distributions of left and right. By default, checks that the supports are identical and throws an error otherwise. Can implement specific fusions for specific distributions.

See also: prod, ProductOf